worker/jobs/downloads: Split file into multiple modules
This makes the file and the test suite for it a little more manageable :)
