+++ src/worker/jobs/downloads/mod.rs
@@ -1,0 +1,5 @@
+mod process_log;
+mod queue;
+
+pub use process_log::ProcessCdnLog;
+pub use queue::ProcessCdnLogQueue;
