+++ src/worker/jobs/downloads/queue/fixtures/empty-event.json
