+++ src/worker/jobs/downloads/queue/mod.rs
@@ -1,0 +1,4 @@
+mod job;
+mod message;
+
+pub use job::ProcessCdnLogQueue;
